Moving cpa.rb to dashboard/lib and renaming CPA -> Cpa
