+++ dashboard/app/controllers/application_controller.rb
@@ -3,7 +3,7 @@ require 'dynamic_config/dcdo'
 require 'dynamic_config/gatekeeper'
 require 'dynamic_config/page_mode'
 require 'cdo/shared_constants'
-require 'cdo/cpa'
+require 'cpa'
 
 class ApplicationController < ActionController::Base
   include LocaleHelper
@@ -330,7 +330,7 @@ class ApplicationController < ActionController::Base
   # are not compliant, then we need to send them to the lockout page.
   def assert_child_account_policy
     # Check that the child account policy is currently enabled.
-    return unless CPA.cpa_experience(request)
+    return unless Cpa.cpa_experience(request)
 
     # Anonymous users are NOT affected by our Child Account Policy
     return unless current_user
