+++ dashboard/app/views/layouts/application.html.haml
@@ -1,5 +1,5 @@
 !!! 5
--require 'cdo/cpa'
+-require 'cpa'
 -require '../shared/middleware/helpers/experiments'
 %html{dir: locale_dir}
   %head
@@ -10,7 +10,7 @@
       window.OFFLINE_SERVICE_WORKER_PATH =
         "/s/express-2021/lessons/1#{webpack_offline_asset_path('js/offline-service-worker.js')}"
     :javascript
-      window.CPA_EXPERIENCE = "#{CPA.cpa_experience(request)}"
+      window.CPA_EXPERIENCE = "#{Cpa.cpa_experience(request)}"
     :javascript
       window.GENDER_FEATURE_ENABLED = "#{experiment_value('gender', request)}"
     - appname = Rails.env.production? ? t(:appname) : "#{t(:appname)} [#{Rails.env}]"
