+++ dashboard/test/controllers/sessions_controller_test.rb
@@ -6,9 +6,9 @@ class SessionsControllerTest < ActionController::TestCase
 
   setup do
     @request.env["devise.mapping"] = Devise.mappings[:user]
-    CPA.stubs(:cpa_experience).
+    Cpa.stubs(:cpa_experience).
       with(any_parameters).
-      returns(CPA::NEW_USER_LOCKOUT)
+      returns(Cpa::NEW_USER_LOCKOUT)
   end
 
   test 'login error derives locale from cdo.locale' do
